Fixed stuff

+++ Frontend/Shared/Pollenalarm.Frontend.Shared/Models/GeoLocation.cs
@@ -8,8 +8,8 @@ namespace Pollenalarm.Frontend.Shared.Models
 {
     public class GeoLocation
     {
-        public long Longitute { get; set; }
-        public long Latitute { get; set; }
+        public double Longitute { get; set; }
+        public double Latitute { get; set; }
         public string Name { get; set; }
         public string Zip { get; set; }
     }
